wip/refactor: frontend online users list
