+++ frontend/src/components/chat/ChatWidget.css
@@ -166,4 +166,3 @@
   background: #f472b6;
 } */
 
-.me
